Unify harness code
