+++ test-harness/tests/concurrent.rs
@@ -10,16 +10,10 @@
 
 use futures::prelude::*;
 use futures::stream::FuturesUnordered;
-use quickcheck::{Arbitrary, Gen, QuickCheck};
-use std::{
-    io,
-    net::{Ipv4Addr, SocketAddr, SocketAddrV4},
-};
+use quickcheck::QuickCheck;
 use test_harness::*;
-use tokio::net::{TcpListener, TcpStream};
-use tokio::{net::TcpSocket, runtime::Runtime, task};
-use tokio_util::compat::{Compat, TokioAsyncReadCompatExt};
-use yamux::{Config, Connection, ConnectionError, Control, Mode, WindowUpdateMode};
+use tokio::{runtime::Runtime, task};
+use yamux::{Config, ConnectionError, Control, WindowUpdateMode};
 
 const PAYLOAD_SIZE: usize = 128 * 1024;
 
@@ -32,7 +26,9 @@ fn concurrent_streams() {
         let n_streams = 1000;
 
         Runtime::new().expect("new runtime").block_on(async move {
-            let (server, client) = connected_peers(tcp_buffer_sizes).await.unwrap();
+            let (server, client) = connected_peers(config(), config(), tcp_buffer_sizes)
+                .await
+                .unwrap();
 
             task::spawn(echo_server(server));
 
@@ -68,73 +64,6 @@ fn concurrent_streams() {
     }
 
     QuickCheck::new().tests(3).quickcheck(prop as fn(_) -> _)
-}
-
-/// Send and receive buffer size for a TCP socket.
-#[derive(Clone, Debug, Copy)]
-struct TcpBufferSizes {
-    send: u32,
-    recv: u32,
-}
-
-impl Arbitrary for TcpBufferSizes {
-    fn arbitrary(g: &mut Gen) -> Self {
-        let send = if bool::arbitrary(g) {
-            16 * 1024
-        } else {
-            32 * 1024
-        };
-
-        // Have receive buffer size be some multiple of send buffer size.
-        let recv = if bool::arbitrary(g) {
-            send * 2
-        } else {
-            send * 4
-        };
-
-        TcpBufferSizes { send, recv }
-    }
-}
-
-async fn connected_peers(
-    buffer_sizes: Option<TcpBufferSizes>,
-) -> io::Result<(Connection<Compat<TcpStream>>, Connection<Compat<TcpStream>>)> {
-    let (listener, addr) = bind(buffer_sizes).await?;
-
-    let server = async {
-        let (stream, _) = listener.accept().await?;
-        Ok(Connection::new(stream.compat(), config(), Mode::Server))
-    };
-    let client = async {
-        let stream = new_socket(buffer_sizes)?.connect(addr).await?;
-
-        Ok(Connection::new(stream.compat(), config(), Mode::Client))
-    };
-
-    futures::future::try_join(server, client).await
-}
-
-async fn bind(buffer_sizes: Option<TcpBufferSizes>) -> io::Result<(TcpListener, SocketAddr)> {
-    let socket = new_socket(buffer_sizes)?;
-    socket.bind(SocketAddr::V4(SocketAddrV4::new(
-        Ipv4Addr::new(127, 0, 0, 1),
-        0,
-    )))?;
-
-    let listener = socket.listen(1024)?;
-    let address = listener.local_addr()?;
-
-    Ok((listener, address))
-}
-
-fn new_socket(buffer_sizes: Option<TcpBufferSizes>) -> io::Result<TcpSocket> {
-    let socket = TcpSocket::new_v4()?;
-    if let Some(size) = buffer_sizes {
-        socket.set_send_buffer_size(size.send)?;
-        socket.set_recv_buffer_size(size.recv)?;
-    }
-
-    Ok(socket)
 }
 
 fn config() -> Config {
